delays processing update
